+++ fabric/src/main/java/de/clickism/clickvillagers/ClickVillagers.java
@@ -7,13 +7,23 @@
 package de.clickism.clickvillagers;
 
 import de.clickism.clickvillagers.callback.*;
+import de.clickism.clickvillagers.util.MessageType;
+import de.clickism.clickvillagers.util.VersionHelper;
+import de.clickism.configured.fabriccommandadapter.FabricCommandAdapter;
+import de.clickism.configured.fabriccommandadapter.command.GetCommand;
+import de.clickism.configured.fabriccommandadapter.command.PathCommand;
+import de.clickism.configured.fabriccommandadapter.command.ReloadCommand;
+import de.clickism.configured.fabriccommandadapter.command.SetCommand;
 import de.clickism.modrinthupdatechecker.ModrinthUpdateChecker;
 import net.fabricmc.api.ModInitializer;
+import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
 import net.fabricmc.fabric.api.event.player.UseBlockCallback;
 import net.fabricmc.fabric.api.event.player.UseEntityCallback;
 import net.fabricmc.fabric.api.networking.v1.ServerPlayConnectionEvents;
 import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.MinecraftVersion;
+import net.minecraft.server.command.CommandManager;
+import net.minecraft.text.Text;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -36,6 +46,28 @@ public class ClickVillagers implements ModInitializer {
             checkUpdates();
             ServerPlayConnectionEvents.JOIN.register(new UpdateNotifier(() -> newerVersion));
         }
+
+        // Register Commands
+        CommandRegistrationCallback.EVENT.register((dispatcher, registryAccess, environment) -> {
+            dispatcher.register(CommandManager.literal("clickvillagers")
+                    .requires(VersionHelper::isOp)
+                    .then(FabricCommandAdapter.ofConfig(CONFIG)
+                            .add(new SetCommand((sender, key, value) -> {
+                                MessageType.CONFIG.send(sender, Text.literal("§aConfig option \"§l" + key + "\" §aset to §l" + value + "."));
+                            }))
+                            .add(new GetCommand((sender, key, value) -> {
+                                MessageType.CONFIG.send(sender, Text.literal("§aConfig option \"§l" + key + "\" §ahas value §l" + value + "."));
+                            }))
+                            .add(new ReloadCommand(sender -> {
+                                MessageType.CONFIG.send(sender, Text.literal("§aReloaded the config file."));
+                            }))
+                            .add(new PathCommand((sender, path) -> {
+                                MessageType.CONFIG.send(sender, Text.literal("§aThe config file is located at: §f" + path));
+                            }))
+                            .buildRoot()
+                    )
+            );
+        });
     }
 
     private void checkUpdates() {
@@ -46,7 +78,7 @@ public class ClickVillagers implements ModInitializer {
         String minecraftVersion = MinecraftVersion.create().name();
         //?} elif >= 1.21.6 {
         /*String minecraftVersion = MinecraftVersion.CURRENT.name();
-        *///?} else
+         *///?} else
         /*String minecraftVersion = MinecraftVersion.CURRENT.getName();*/
         new ModrinthUpdateChecker(MOD_ID, "fabric", minecraftVersion).checkVersion(version -> {
             if (modVersion == null || ModrinthUpdateChecker.getRawVersion(modVersion).equals(version)) {