+++ fabric/src/main/java/de/clickism/clickvillagers/util/VersionHelper.java
@@ -22,6 +22,7 @@ import net.minecraft.recipe.Recipe;
 import net.minecraft.server.MinecraftServer;
 //? if >=1.21.9
 import net.minecraft.server.PlayerConfigEntry;
+import net.minecraft.server.command.ServerCommandSource;
 import net.minecraft.sound.SoundCategory;
 import net.minecraft.sound.SoundEvent;
 import net.minecraft.util.UserCache;
@@ -35,7 +36,16 @@ import java.util.UUID;
 public class VersionHelper {
     public static void playSound(PlayerEntity player, SoundEvent soundEvent, SoundCategory category, float volume, float pitch) {
         //? if >=1.21.11 {
-        player.playSound(soundEvent, volume, pitch);
+        player.getEntityWorld().playSound(
+                player,
+                player.getX(),
+                player.getY(),
+                player.getZ(),
+                soundEvent,
+                category,
+                volume,
+                pitch
+        );
         //?} elif >=1.20.5 {
         /*player.playSoundToPlayer(soundEvent, category, volume, pitch);
          *///?} else
@@ -102,5 +112,14 @@ public class VersionHelper {
                || perms.hasPermission(DefaultPermissions.OWNERS);
         //?} else
         /*return player.hasPermissionLevel(3);*/
+    }
+
+    public static boolean isOp(ServerCommandSource source) {
+        //? if >=1.21.11 {
+        var perms = source.getPermissions();
+        return perms.hasPermission(DefaultPermissions.ADMINS)
+               || perms.hasPermission(DefaultPermissions.OWNERS);
+        //?} else
+        /*return source.hasPermissionLevel(3);*/
     }
 }