+++ fabric/CHANGELOG.md
@@ -1,3 +1,6 @@
-- For 1.21.11: Downgraded Fabric loader dependency to `0.18.0`.
-- You can now place duplicated picked up villager items. (Matches behavior of the Paper version)
-  - Duplicated villagers will be assigned new UUIDs.
+- Fixed: Command/interaction sounds wouldn't play on 1.21.11.
+- **Updated to Configured 0.3**: Internal refactor and added command interface for configuration options!
+  - `/clickvillagers config set <option> <value>`: Sets a configuration option.
+  - `/clickvillagers config get <option>`: Gets the value of a configuration option.
+  - `/clickvillagers config path`: Prints the path to the **configuration file**.
+  - `/clickvillagers config reload`: Reloads the configuration file.